Fixes incorrect assumptions in block type compute
We allow tensors where different dimensions have different nesting
levels.  For example a vector of matrices, as in:
tensor[springs](tensor[2,2](float)).  When reasoning about it at the
top level only it is a vector we can insert matrices into, when
reasoning about the whole thing it is a matrix.