Remove remove llvm optimization from debug builds
This is because on my system (and Shin’s) there’s some llvm bug that
causes programs to segfault when linking in the ipo library. We’ll
change llvm later, but for now this is a workaround.

The bug might be because we don’t create correct verified llvm code in
some cases.