Adds localize-temps pass to GPU, packs tensor data into struct
Temporaries cannot be allocated at the global level if they are
declared inside a kernel, because parallel iterations of the
kernel will conflict. We instead move these to mallocs within
the kernel itself. This is not appropriate for system-sized
objects, so the mid-end should take care to lower loops
appropriately to avoid this case.

Tensor data is now packed into a struct to provide some extra
information to the GPUFunction object, which the LLVM backend
instead defers to runtime calculation in compiled code.